+++ openbrokerapi/api.py
@@ -298,6 +298,16 @@ def get_blueprint(service_broker: ServiceBroker,
         result = service_broker.last_operation(instance_id, operation_data)
         return to_json_response(LastOperationResponse(result.state, result.description)), HTTPStatus.OK
 
+    @openbroker.route("/v2/service_instances/<instance_id>/service_bindings/<binding_id>/last_operation", methods=['GET'])
+    def last_binding_operation(instance_id, binding_id):
+        # TODO: forward them
+        # service_id = request.args.get("service_id", None)
+        # plan_id = request.args.get("plan_id", None)
+
+        operation_data = request.args.get("operation", None)
+        result = service_broker.last_binding_operation(instance_id, binding_id, operation_data)
+        return to_json_response(LastOperationResponse(result.state, result.description)), HTTPStatus.OK
+
     return openbroker